requirements.txt
python-telegram-bot==20.7
matplotlib==3.7.1
numpy==1.24.3